+++ doc/changelog.rst
@@ -7,6 +7,9 @@ Version 1.0.1 (development version)
 
 * Fix Trollius issue #9: @trollius.coroutine now works on callbable objects
   (without ``__name__`` attribute), not only on functions.
+* Fix Trollius issue #13: asyncio futures are now accepted in all functions:
+  as_completed(), async(), @coroutine, gather(), run_until_complete(),
+  wrap_future().
 
 
 2014-07-21: Version 1.0